% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ConfigFile}
\alias{ConfigFile}
\title{Google Cloud Deployment Manager API Objects 
Declares, configures, and deploys complex solutions on Google Cloud Platform.}
\usage{
ConfigFile(content = NULL)
}
\arguments{
\item{content}{The contents of the file}
}
\value{
ConfigFile object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:06:05
filename: /Users/mark/dev/R/autoGoogleAPI/googledeploymentmanagerv2.auto/R/deploymentmanager_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ConfigFile Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}

