% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Deployment}
\alias{Deployment}
\title{Deployment Object}
\usage{
Deployment(description = NULL, fingerprint = NULL, id = NULL,
  insertTime = NULL, labels = NULL, manifest = NULL, name = NULL,
  operation = NULL, selfLink = NULL, target = NULL, update = NULL)
}
\arguments{
\item{description}{An optional user-provided description of the deployment}

\item{fingerprint}{Provides a fingerprint to use in requests to modify a deployment, such as update(), stop(), and cancelPreview() requests}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{insertTime}{[Output Only] Timestamp when the deployment was created, in RFC3339 text format}

\item{labels}{Map of labels; provided by the client when the resource is created or updated}

\item{manifest}{[Output Only] URL of the manifest representing the last manifest that was successfully deployed}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{operation}{[Output Only] The Operation that most recently ran, or is currently running, on this deployment}

\item{selfLink}{[Output Only] Self link for the deployment}

\item{target}{[Input Only] The parameters that define your deployment, including the deployment configuration and relevant templates}

\item{update}{[Output Only] If Deployment Manager is currently updating or previewing an update to this deployment, the updated configuration appears here}
}
\value{
Deployment object
}
\description{
Deployment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
\seealso{
Other Deployment functions: \code{\link{deployments.insert}},
  \code{\link{deployments.patch}},
  \code{\link{deployments.update}}
}

