% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_functions.R
\name{deployments.insert}
\alias{deployments.insert}
\title{Creates a deployment and all of the resources described by the deployment manifest.}
\usage{
deployments.insert(Deployment, project, preview = NULL)
}
\arguments{
\item{Deployment}{The \link{Deployment} object to pass to this method}

\item{project}{The project ID for this request}

\item{preview}{If set to true, creates a deployment and creates 'shell' resources but does not actually instantiate these resources}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/}{Google Documentation}

Other Deployment functions: \code{\link{Deployment}},
  \code{\link{deployments.patch}},
  \code{\link{deployments.update}}
}

