% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_functions.R
\name{deployments.stop}
\alias{deployments.stop}
\title{Stops an ongoing operation. This does not roll back any work that has already been completed, but prevents any new work from being started.}
\usage{
deployments.stop(DeploymentsStopRequest, project, deployment)
}
\arguments{
\item{DeploymentsStopRequest}{The \link{DeploymentsStopRequest} object to pass to this method}

\item{project}{The project ID for this request}

\item{deployment}{The name of the deployment for this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/ndev.cloudman
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/ndev.cloudman)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/}{Google Documentation}

Other DeploymentsStopRequest functions: \code{\link{DeploymentsStopRequest}}
}

