% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{DeploymentsCancelPreviewRequest}
\alias{DeploymentsCancelPreviewRequest}
\title{DeploymentsCancelPreviewRequest Object}
\usage{
DeploymentsCancelPreviewRequest(fingerprint = NULL)
}
\arguments{
\item{fingerprint}{Specifies a fingerprint for cancelPreview() requests}
}
\value{
DeploymentsCancelPreviewRequest object
}
\description{
DeploymentsCancelPreviewRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}
\seealso{
Other DeploymentsCancelPreviewRequest functions: \code{\link{deployments.cancelPreview}}
}

