% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ImportFile}
\alias{ImportFile}
\title{ImportFile Object}
\usage{
ImportFile(content = NULL, name = NULL)
}
\arguments{
\item{content}{The contents of the file}

\item{name}{The name of the file}
}
\value{
ImportFile object
}
\description{
ImportFile Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}

