% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{LogConfigCounterOptions}
\alias{LogConfigCounterOptions}
\title{LogConfigCounterOptions Object}
\usage{
LogConfigCounterOptions(field = NULL, metric = NULL)
}
\arguments{
\item{field}{The field value to attribute}

\item{metric}{The metric to update}
}
\value{
LogConfigCounterOptions object
}
\description{
LogConfigCounterOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for counters
}
