% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ManifestsListResponse}
\alias{ManifestsListResponse}
\title{ManifestsListResponse Object}
\usage{
ManifestsListResponse(manifests = NULL, nextPageToken = NULL)
}
\arguments{
\item{manifests}{[Output Only] Manifests contained in this list response}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}
}
\value{
ManifestsListResponse object
}
\description{
ManifestsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A response containing a partial list of manifests and a page token used to build the next request if the request has been truncated.
}

