% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Resource.warnings.data}
\alias{Resource.warnings.data}
\title{Resource.warnings.data Object}
\usage{
Resource.warnings.data()
}
\value{
Resource.warnings.data object
}
\description{
Resource.warnings.data Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Resource functions: \code{\link{Resource.warnings}},
  \code{\link{Resource}}
}

