% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{Resource.warnings}
\alias{Resource.warnings}
\title{Resource.warnings Object}
\usage{
Resource.warnings(Resource.warnings.data = NULL)
}
\arguments{
\item{Resource.warnings.data}{The \link{Resource.warnings.data} object or list of objects}
}
\value{
Resource.warnings object
}
\description{
Resource.warnings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Resource functions: \code{\link{Resource.warnings.data}},
  \code{\link{Resource}}
}

