% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ResourceAccessControl}
\alias{ResourceAccessControl}
\title{ResourceAccessControl Object}
\usage{
ResourceAccessControl(gcpIamPolicy = NULL)
}
\arguments{
\item{gcpIamPolicy}{The GCP IAM Policy to set on the resource}
}
\value{
ResourceAccessControl object
}
\description{
ResourceAccessControl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The access controls set on the resource.
}
