% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ResourceUpdate.error}
\alias{ResourceUpdate.error}
\title{ResourceUpdate.error Object}
\usage{
ResourceUpdate.error(ResourceUpdate.error.errors = NULL, errors = NULL)
}
\arguments{
\item{ResourceUpdate.error.errors}{The \link{ResourceUpdate.error.errors} object or list of objects}

\item{errors}{[Output Only] The array of errors encountered while processing this operation}
}
\value{
ResourceUpdate.error object
}
\description{
ResourceUpdate.error Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Output Only] If errors are generated during update of the resource, this field will be populated.
}
\seealso{
Other ResourceUpdate functions: \code{\link{ResourceUpdate.error.errors}},
  \code{\link{ResourceUpdate.warnings.data}},
  \code{\link{ResourceUpdate.warnings}},
  \code{\link{ResourceUpdate}}
}

