% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{ResourceUpdate.warnings}
\alias{ResourceUpdate.warnings}
\title{ResourceUpdate.warnings Object}
\usage{
ResourceUpdate.warnings(ResourceUpdate.warnings.data = NULL)
}
\arguments{
\item{ResourceUpdate.warnings.data}{The \link{ResourceUpdate.warnings.data} object or list of objects}
}
\value{
ResourceUpdate.warnings object
}
\description{
ResourceUpdate.warnings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ResourceUpdate functions: \code{\link{ResourceUpdate.error.errors}},
  \code{\link{ResourceUpdate.error}},
  \code{\link{ResourceUpdate.warnings.data}},
  \code{\link{ResourceUpdate}}
}

