% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploymentmanager_objects.R
\name{TargetConfiguration}
\alias{TargetConfiguration}
\title{TargetConfiguration Object}
\usage{
TargetConfiguration(config = NULL, imports = NULL)
}
\arguments{
\item{config}{The configuration to use for this deployment}

\item{imports}{Specifies any files to import for this configuration}
}
\value{
TargetConfiguration object
}
\description{
TargetConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
}

